refactor: useSquadInvitation
merge changes with upstream hook
